re-define `.arity` field for Function values as an `int8` (the `-1` is needed as a default *not found* alias + 127 possible arguments for a function is... I guess more than enough lol)

+++ src/vm/values/types.nim
@@ -148,7 +148,7 @@ type
                 discard
 
     VFunction* = ref object
-        arity*  : int
+        arity*  : int8
         case fnKind*: FunctionKind:
             of UserFunction:
                 # TODO(VM/values/types) merge Function `params` and `args` into one field?
